+++ ethereum/contracts/interfaces/ISwap.sol
@@ -7,8 +7,8 @@ interface ISwap is IProviderRegistry {
     struct SwapParams {
         bytes4 providerID;
         address[] path;
-        uint amountIn;
-        uint minAmountOut;
+        uint256 amountIn;
+        uint256 minAmountOut;
         address recipient;
         bytes data;
     }
@@ -18,32 +18,32 @@ interface ISwap is IProviderRegistry {
     function getAmountIn(
         bytes4 providerID,
         address[] memory path,
-        uint amountOut
-    ) external returns (uint amountIn, bytes memory swapData);
+        uint256 amountOut
+    ) external returns (uint256 amountIn, bytes memory swapData);
 
     // @dev calculates the maximum tokens can be transferred for the amountIn
     // @return swapData the data to be passed to the swap
     function getAmountOut(
         bytes4 providerID,
         address[] memory path,
-        uint amountIn
-    ) external returns (uint amountOut, bytes memory swapData);
+        uint256 amountIn
+    ) external returns (uint256 amountOut, bytes memory swapData);
 
     // @dev view only version of getAmountIn
     function getAmountInView(
         bytes4 providerID,
         address[] memory path,
-        uint amountOut
-    ) external view returns (uint amountIn, bytes memory swapData);
+        uint256 amountOut
+    ) external view returns (uint256 amountIn, bytes memory swapData);
 
     // @dev view only version of getAmountOut
     function getAmountOutView(
         bytes4 providerID,
         address[] memory path,
-        uint amountIn
-    ) external view returns (uint amountOut, bytes memory swapData);
+        uint256 amountIn
+    ) external view returns (uint256 amountOut, bytes memory swapData);
 
     // @dev calls swap adapter to fulfill the exchange
     // @return amountOut the amount of tokens transferred out, guarantee amountOut >= params.minAmountOut
-    function swap(SwapParams calldata params) external payable returns (uint amountOut);
+    function swap(SwapParams calldata params) external payable returns (uint256 amountOut);
 }
