+++ ethereum/contracts/interfaces/swaps/curve/IAddressProvider.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.11;
 
 interface IAddressProvider {
-    function get_address(uint id) external view returns (address addr);
+    function get_address(uint256 id) external view returns (address addr);
 
-    function max_id() external view returns (uint id);
+    function max_id() external view returns (uint256 id);
 }
