+++ ethereum/contracts/interfaces/swaps/curve/IExchange.sol
@@ -2,24 +2,24 @@
 pragma solidity ^0.8.11;
 
 interface IExchange {
-    function get_best_rate(address from, address to, uint amount) external view returns (address, uint);
+    function get_best_rate(address from, address to, uint256 amount) external view returns (address, uint);
 
-    function get_exchange_amount(address pool, address from, address to, uint amount) external view returns (uint);
+    function get_exchange_amount(address pool, address from, address to, uint256 amount) external view returns (uint);
 
     function exchange(
         address pool,
         address from,
         address to,
-        uint amount,
-        uint expected,
+        uint256 amount,
+        uint256 expected,
         address receiver
-    ) external payable returns (uint);
+    ) external payable returns (uint256);
 
     function exchange_with_best_rate(
         address from,
         address to,
-        uint amount,
-        uint expected,
+        uint256 amount,
+        uint256 expected,
         address receiver
-    ) external payable returns (uint);
+    ) external payable returns (uint256);
 }
